完成院校部分的分页，根据用户点击学校详情使用幻灯片实现智能推荐，修改后台幻灯片的删除部分，使得删除的同时删除磁盘文件，修复院校搜索的bug

+++ sais-service/src/main/java/com/sais/saisservice/SlideService.java
@@ -19,6 +19,8 @@ import java.util.List;
 @Service
 public class SlideService {
     private SlideMapper slideMapper;
+    @Value("${file.collegeImagesUploadPath}")
+    private String collegeImagesUploadPath;
 
     @Autowired
     public SlideService(SlideMapper slideMapper){
@@ -30,15 +32,26 @@ public class SlideService {
     }
 
     public int delete(int id){
-        return slideMapper.delete(id);
+        Slide slide=slideMapper.selectId(id);
+        File file = new File(collegeImagesUploadPath+slide.getPicture());
+        if (file.isFile() && file.exists()) {
+            if(file.delete()){
+                return slideMapper.delete(id);
+            }
+        }
+        return 0;
     }
 
-    public int delAll(ArrayList<String> list){
-        return slideMapper.delAll(list);
+    public boolean delAll(ArrayList<Integer> list){
+        for(Integer id : list){
+            if(delete(id)==0){
+                return false;
+            }
+        }
+        return true;
     }
 
-    public ArrayList<Slide> lists(Integer pageNum,Integer pageSize){
-        PageHelper.startPage(pageNum, pageSize);
+    public ArrayList<Slide> lists(){
         return slideMapper.lists();
     }
